cmake: Fix C++11 feature detection for pre-3.1 CMake
With older versions of Cmake, pre-3.1, the feature detection macro,
target_compile_features, is not available. As a side effect, distributions
like Debian 8 cannot compile Chigraph due to their old versions of CMake.

This commit adds a features detection scheme that supports the commonly used
C++11 features and works with CMake versions as far back as 2.8.3. Additionally,
this variant is conditionally used and target_compile_features is favored when
available.

ISSUE: #75
TEST: Configuration succeeds on Debian 8 (CMAKE 3.0.2)

Tested-by: Aun-Ali Zaidi <admin@kodeit.net>
Signed-off-by: Aun-Ali Zaidi <admin@kodeit.net>